Remove Microsoft Word comments
When users add comments to Word 2016 documents and then copy that
content the word document contains the comments that are output.

This commit removes these comments. It does this by matching the
particular element and class name combinations produced by Microsoft
Word.

This is likely very delicate as it requires MS Office to consistently
produce the same code even in future versions.

The HTML snippet to test this was rather big for testing an individual
scenario so I decided to create a fixture for this and include the full
HTML that is output. After chatting with Alex a directory of
__fixtures__ was deemed a common JS library convention.